Fix integer sanitizer warning in hash.c
BUG=aomedia:2357

+++ av1/encoder/hash.c
@@ -15,8 +15,8 @@ static void crc_calculator_process_data(CRC_CALCULATOR *p_crc_calculator,
                                         uint8_t *pData, uint32_t dataLength) {
   for (uint32_t i = 0; i < dataLength; i++) {
     const uint8_t index =
-        (p_crc_calculator->remainder >> (p_crc_calculator->bits - 8)) ^
-        pData[i];
+        (uint8_t)((p_crc_calculator->remainder >> (p_crc_calculator->bits - 8)) ^
+        pData[i]);
     p_crc_calculator->remainder <<= 8;
     p_crc_calculator->remainder ^= p_crc_calculator->table[index];
   }
